Refactor shortcut/donate code to avoid duplication
